Upgrade testing to avoid Gestalt.
strerror, our test-victim function that we override, returns a slightly-different result on 10.6 and 10.7+.

We used to call Gestalt to decide which we should expect, but Gestalt is deprecated. Apparently the current "best" solution is to read SystemVersion.plist  (ref https://twitter.com/rentzsch/status/373203814993494016) but that's a lot of code to write just to decide on which string to expect. So this code now will accept either.

But the assert* code was hard to follow with the strcmp() == 0, so I upgraded that yak as well.